Delete Singleton as it makes unit testing difficult. Refactor Config instantiation to use factory method.
This change replaces the direct instantiation of the Config class with a factory method from the newly created ConfigFactory class. This is done to hide the complexity of Config object creation and to adhere to the principle of separating the responsibility of object creation from its usage. This change includes the deletion of the constructor call in main.cpp, and the addition of an include directive for ConfigFactory.hpp.