Refactor environment variables retrieval and type changes
The changes introduce separate methods for retrieving environment variables in Windows and non-Windows systems. Also, the type of parameter 'path' for 'resolvePath' function in IFileSystem and its implementations has been changed from 'std::string_view' to 'std::string'. This was done to increase consistency across platforms. The optimization level flag for MSVC in CMakeLists.txt was adjusted to '/Ox'.